+++ lib/cli/versions.json
@@ -1,17 +1,55 @@
 {
+  "@storybook/addon-a11y": "6.0.0-rc.3",
+  "@storybook/addon-actions": "6.0.0-rc.3",
+  "@storybook/addon-backgrounds": "6.0.0-rc.3",
+  "@storybook/addon-controls": "6.0.0-rc.3",
+  "@storybook/addon-cssresources": "6.0.0-rc.3",
+  "@storybook/addon-design-assets": "6.0.0-rc.3",
+  "@storybook/addon-docs": "6.0.0-rc.3",
+  "@storybook/addon-essentials": "6.0.0-rc.3",
+  "@storybook/addon-events": "6.0.0-rc.3",
+  "@storybook/addon-google-analytics": "6.0.0-rc.3",
+  "@storybook/addon-graphql": "6.0.0-rc.3",
+  "@storybook/addon-jest": "6.0.0-rc.3",
+  "@storybook/addon-knobs": "6.0.0-rc.3",
+  "@storybook/addon-links": "6.0.0-rc.3",
+  "@storybook/addon-queryparams": "6.0.0-rc.3",
+  "@storybook/addon-storyshots": "6.0.0-rc.3",
+  "@storybook/addon-storyshots-puppeteer": "6.0.0-rc.3",
+  "@storybook/addon-storysource": "6.0.0-rc.3",
+  "@storybook/addon-toolbars": "6.0.0-rc.3",
+  "@storybook/addon-viewport": "6.0.0-rc.3",
+  "@storybook/addons": "6.0.0-rc.3",
   "@storybook/angular": "6.0.0-rc.3",
+  "@storybook/api": "6.0.0-rc.3",
   "@storybook/aurelia": "6.0.0-rc.3",
+  "@storybook/channel-postmessage": "6.0.0-rc.3",
+  "@storybook/channel-websocket": "6.0.0-rc.3",
+  "@storybook/channels": "6.0.0-rc.3",
+  "@storybook/cli": "6.0.0-rc.3",
+  "@storybook/client-api": "6.0.0-rc.3",
+  "@storybook/client-logger": "6.0.0-rc.3",
+  "@storybook/codemod": "6.0.0-rc.3",
+  "@storybook/components": "6.0.0-rc.3",
+  "@storybook/core": "6.0.0-rc.3",
+  "@storybook/core-events": "6.0.0-rc.3",
   "@storybook/ember": "6.0.0-rc.3",
   "@storybook/html": "6.0.0-rc.3",
   "@storybook/marionette": "6.0.0-rc.3",
   "@storybook/marko": "6.0.0-rc.3",
   "@storybook/mithril": "6.0.0-rc.3",
+  "@storybook/node-logger": "6.0.0-rc.3",
+  "@storybook/postinstall": "6.0.0-rc.3",
   "@storybook/preact": "6.0.0-rc.3",
   "@storybook/rax": "6.0.0-rc.3",
   "@storybook/react": "6.0.0-rc.3",
   "@storybook/riot": "6.0.0-rc.3",
+  "@storybook/router": "6.0.0-rc.3",
   "@storybook/server": "6.0.0-rc.3",
+  "@storybook/source-loader": "6.0.0-rc.3",
   "@storybook/svelte": "6.0.0-rc.3",
+  "@storybook/theming": "6.0.0-rc.3",
+  "@storybook/ui": "6.0.0-rc.3",
   "@storybook/vue": "6.0.0-rc.3",
   "@storybook/web-components": "6.0.0-rc.3"
 }
